first cut of connection helper
git-svn-id: https://viaduct.svn.beanstalkapp.com/code/trunk/viaduct@89 c97fca75-8364-49ce-9f5b-20eef4eea7b5

+++ src/ngx_http_viaduct/ngx_http_viaduct_module.c
@@ -344,7 +344,7 @@ ngx_http_viaduct_create_loc_conf(ngx_conf_t *cf)
 static void 
 copy_value(char *dest, char *src, int sz)
 {
-   if (strlen(src) < sz) strcpy(dest, src);
+   if (strlen(src) < (unsigned int) sz) strcpy(dest, src);
    else {
       strncpy(dest, src, sz - 1);
       dest[sz-1]='\0';
